--- app/controllers/api/v1/city_details_controller.rb
@@ -1,6 +1,0 @@
-class Api::V1::CityDetailsController < ApplicationController
-def index
-    city = CityDetailsFacade.city_data(params[:city], params[:state_initials])
-    render json: CityDetailsSerializer.new(city)
-  end
-end
