+++ app/poros/city.rb
@@ -3,13 +3,24 @@ class City
               :total_score,
               :city,
               :state,
-              :picture
+              :picture,
+              :categories,
+              :id
 
   def initialize(data, photo_data)
+    @id = nil
     @summary = data[:_embedded][:"city:search-results"][0][:_embedded][:"city:item"][:_embedded][:"city:urban_area"][:_embedded][:"ua:scores"][:summary]
     @total_score = data[:_embedded][:"city:search-results"][0][:_embedded][:"city:item"][:_embedded][:"city:urban_area"][:_embedded][:"ua:scores"][:teleport_city_score].round(2)
     @city = data[:_embedded][:"city:search-results"][0][:_embedded][:"city:item"][:_embedded][:"city:urban_area"][:name]
     @state = data[:_embedded][:"city:search-results"][0][:_embedded][:"city:item"][:_embedded][:"city:urban_area"][:_links][:"ua:admin1-divisions"][0][:name]
     @picture = photo_data[:photos][0][:image][:web]
+    @categories = data[:_embedded][:"city:search-results"][0][:_embedded][:"city:item"][:_embedded][:"city:urban_area"][:_embedded][:"ua:scores"][:categories]
+  end
+
+  def categories_hash_array
+    @categories.map do |category|
+      category.delete(:color)
+      category
+    end
   end
 end
