--- app/serializers/city_detail_serializer.rb
@@ -1,4 +1,0 @@
-class CityDetailSerializer
-  include JSONAPI::Serializer
-  attributes :categories
-end
