+++ app/serializers/city_serializer.rb
@@ -1,4 +1,4 @@
 class CitySerializer
   include JSONAPI::Serializer
-  attributes :summary, :total_score, :city, :state, :picture
+  attributes :summary, :total_score, :city, :state, :picture, :categories
 end
