--- spec/api/v1/city_details_index_spec.rb
@@ -1,15 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe 'index' do
-  before :each do
-
-    create
-    visit
-
-    end
-
-  describe '' do
-    describe '.' do
-    end
-
-  end
